+++ lib/td_audit/notifications/notification_loader.ex
@@ -11,7 +11,7 @@ defmodule TdAudit.NotificationLoader do
   @notification_load_frequency Application.get_env(
                                  :td_audit,
                                  :notification_load_frequency
-                               )
+                               )[:events]
 
   def start_link(name \\ nil) do
     GenServer.start_link(__MODULE__, nil, name: name)
@@ -19,35 +19,45 @@ defmodule TdAudit.NotificationLoader do
 
   @impl true
   def init(state) do
-    schedule_work()
+    schedule_work(:create_comment)
+    schedule_work(:failed_rule_results)
     {:ok, state}
   end
 
   @impl true
-  def handle_info(:work, state) do
-    execute_notifications_dispatcher_for_event(%{event: "create_comment"})
-    schedule_work()
+  def handle_info(:failed_rule_results, state) do
+    execute_notifications_dispatcher_for_event(%{event: "failed_rule_results"})
+    schedule_work(:failed_rule_results)
     {:noreply, state}
   end
 
-  defp schedule_work do
-    Process.send_after(self(), :work, @notification_load_frequency)
+  @impl true
+  def handle_info(:create_comment, state) do
+    execute_notifications_dispatcher_for_event(%{event: "create_comment"})
+    schedule_work(:create_comment)
+    {:noreply, state}
+  end
+
+  defp schedule_work(:create_comment) do
+    span = Map.get(@notification_load_frequency, :create_comment)
+    Process.send_after(self(), :create_comment, span)
+  end
+
+  defp schedule_work(:failed_rule_results) do
+    span = Map.get(@notification_load_frequency, :failed_rule_results)
+    Process.send_after(self(), :failed_rule_results, span)
   end
 
   defp execute_notifications_dispatcher_for_event(%{event: "create_comment"} = params) do
-    active =
-      case NotificationsSystem.get_configuration_by_filter(params) do
-        nil ->
-          nil
+    params
+    |> active_configuration?()
+    |> dispatch_notification(%{event: "create_comment"})
+  end
 
-        configuration ->
-          configuration
-          |> Map.fetch!(:settings)
-          |> Map.fetch!("generate_notification")
-          |> Map.fetch!("active")
-      end
-
-    dispatch_notification(active, %{event: "create_comment"})
+  defp execute_notifications_dispatcher_for_event(%{event: "failed_rule_results"} = params) do
+    params
+    |> active_configuration?()
+    |> dispatch_notification(%{event: "failed_rule_results"})
   end
 
   defp dispatch_notification(true, %{event: "create_comment"}) do
@@ -56,7 +66,26 @@ defmodule TdAudit.NotificationLoader do
     )
   end
 
-  defp dispatch_notification(_, %{event: "create_comment"}) do
-    Logger.info("Inactive configuration for event create_comment")
+  defp dispatch_notification(true, %{event: "failed_rule_results"}) do
+    NotificationDispatcher.dispatch_notification(
+      {:dispatch_on_failed_results, "failed_rule_results"}
+    )
+  end
+
+  defp dispatch_notification(_, %{event: event}) do
+    Logger.info("Inactive configuration for event: #{event}")
+  end
+
+  defp active_configuration?(params) do
+    case NotificationsSystem.get_configuration_by_filter(params) do
+      nil ->
+        nil
+
+      configuration ->
+        configuration
+        |> Map.fetch!(:settings)
+        |> Map.fetch!("generate_notification")
+        |> Map.fetch!("active")
+    end
   end
 end