+++ lib/td_audit/subscriptions/subscription_event_processor.ex
@@ -7,17 +7,36 @@ defmodule TdAudit.SubscriptionEventProcessor do
   alias TdCache.UserCache
 
   def process_event(%{"event" => "create_concept_draft"} = event_params) do
-    configuration =
+    configurations =
       Map.new()
       |> Map.put("event", "create_concept_draft")
-      |> NotificationsSystem.get_configuration_by_filter()
+      |> NotificationsSystem.get_configurations_by_filter()
 
-    process_event_from_configuration(configuration, event_params)
+    process_event(configurations, event_params)
+  end
+
+  def process_event(%{"event" => "update_concept_draft"} = event_params) do
+    configurations =
+      Map.new()
+      |> Map.put("event", "create_concept_draft")
+      |> NotificationsSystem.get_configurations_by_filter()
+
+    events =
+      configurations
+      |> Enum.filter(&changed_involved_roles?(&1, event_params))
+      |> Enum.map(&target_event(&1))
+
+    Map.new()
+    |> Map.put("resource_id", Map.get(event_params, "resource_id"))
+    |> Map.put("resource_type", "business_concept")
+    |> Map.put("event", events)
+    |> delete_subscriptions()
+
+    process_event(configurations, event_params)
   end
 
   def process_event(%{"event" => "delete_concept_draft"} = event_params) do
     %{}
-    |> Map.put("event", "create_comment")
     |> Map.put("resource_type", "business_concept")
     |> Map.merge(Map.take(event_params, ["resource_id"]))
     |> delete_subscriptions()
@@ -31,46 +50,91 @@ defmodule TdAudit.SubscriptionEventProcessor do
     )
   end
 
-  defp process_event_from_configuration(nil, _event_params) do
+  defp process_event([], _event_params) do
     Logger.info("No subscription configuration found for event create_concept_draft")
   end
 
-  defp process_event_from_configuration(configuration, event_params) do
-    involved_roles =
-      configuration
-      |> involved_roles_from_configuration()
+  defp process_event(configurations, event_params) when is_list(configurations) do
+    Enum.map(configurations, &process_event_for_configuration(&1, event_params))
+  end
 
-    involved_params =
+  defp process_event_for_configuration(configuration, event_params) do
+    involved_roles = involved_roles(configuration)
+    target_event = target_event(configuration)
+
+    params =
       event_params
       |> Map.take(["payload", "resource_id"])
       |> Map.put("resource_type", "business_concept")
+      |> Map.put("target_event", target_event)
 
     create_subscription_for_roles(
-      involved_params,
+      params,
       involved_roles
     )
   end
 
-  defp involved_roles_from_configuration(%Configuration{settings: settings}) do
+  defp changed_involved_roles?(configuration, event_params) do
+    roles = involved_roles(configuration)
+
+    keys =
+      event_params
+      |> Map.get("payload", %{})
+      |> Map.get("content", %{})
+      |> Map.get("changed", %{})
+      |> Map.keys()
+
+    Enum.any?(keys, &(&1 in roles))
+  end
+
+  defp involved_roles(%Configuration{settings: settings}) do
     settings
     |> Map.fetch!("generate_subscription")
     |> Map.fetch!("roles")
+  end
+
+  defp target_event(%Configuration{settings: settings}) do
+    settings
+    |> Map.fetch!("generate_subscription")
+    |> Map.get("target_event")
+  end
+
+  defp create_subscription_for_roles(
+         %{
+           "payload" => %{"content" => %{"changed" => content}},
+           "resource_id" => resource_id,
+           "resource_type" => resource_type,
+           "target_event" => target_event
+         },
+         involved_roles
+       ) do
+    resource_id
+    |> build_non_user_params(resource_type, target_event)
+    |> create_subscription(involved_roles, content)
   end
 
   defp create_subscription_for_roles(
          %{
            "payload" => %{"content" => content},
            "resource_id" => resource_id,
-           "resource_type" => resource_type
+           "resource_type" => resource_type,
+           "target_event" => target_event
          },
          involved_roles
        ) do
-    non_user_params =
-      %{}
-      |> Map.put("resource_id", resource_id)
-      |> Map.put("resource_type", resource_type)
-      |> Map.put("event", "create_comment")
+    resource_id
+    |> build_non_user_params(resource_type, target_event)
+    |> create_subscription(involved_roles, content)
+  end
 
+  defp build_non_user_params(resource_id, resource_type, target_event) do
+    %{}
+    |> Map.put("resource_id", resource_id)
+    |> Map.put("resource_type", resource_type)
+    |> Map.put("event", target_event || "create_comment")
+  end
+
+  defp create_subscription(non_user_params, involved_roles, content) do
     involved_roles
     |> Enum.map(&Map.get(content, &1))
     |> Enum.filter(&(&1 != nil && is_binary(&1)))