+++ test/td_audit/notifications/notification_dispatcher_test.exs
@@ -9,6 +9,8 @@ defmodule TdAudit.NotificationDispatcherTest do
   alias TdAudit.NotificationDispatcher
   alias TdAudit.Subscriptions
   alias TdCache.ConceptCache
+  alias TdCache.RuleCache
+  alias TdCache.RuleResultCache
   alias TdCache.UserCache
 
   describe "notification_dispatcher" do
@@ -30,9 +32,33 @@ defmodule TdAudit.NotificationDispatcherTest do
       "name" => "BC Name 4",
       "business_concept_version_id" => 1
     }
+    @rule %{
+      id: 1000,
+      active: true,
+      name: "Rule Name",
+      updated_at: DateTime.utc_now(),
+      business_concept_id: @bc_4["id"],
+      minimum: 100
+    }
+    @resut_1 %{
+      id: 1000,
+      date: DateTime.utc_now(),
+      implementation_key: "Key 1",
+      result: 80,
+      rule_id: @rule.id
+    }
+    @resut_2 %{
+      id: 2000,
+      date: DateTime.utc_now(),
+      implementation_key: "Key 2",
+      result: 75,
+      rule_id: @rule.id
+    }
 
     @user_list [@user_1]
     @bc_list [@bc_1, @bc_4]
+    @rules [@rule]
+    @results [@resut_1, @resut_2]
 
     defp list_of_events_to_dipatch do
       event_1 = %{
@@ -98,7 +124,7 @@ defmodule TdAudit.NotificationDispatcherTest do
       [event_1, event_2, event_3, event_4]
     end
 
-    defp list_of_subscriptions do
+    defp list_of_subscriptions(:email_on_comment) do
       # This subscription is one to be sent resource_id => 1
       subscription_1 = %{
         event: "create_comment",
@@ -136,6 +162,26 @@ defmodule TdAudit.NotificationDispatcherTest do
 
       [subscription_1, subscription_2, subscription_3, subscription_4]
     end
+
+    defp list_of_subscriptions(:failed_rule_results) do
+      subscription_1 = %{
+        event: "failed_rule_results",
+        user_email: "mymail1@foo.bar",
+        resource_type: "business_concept",
+        last_consumed_event: DateTime.utc_now(),
+        resource_id: @bc_4["id"]
+      }
+
+      subscription_2 = %{
+        event: "failed_rule_results",
+        user_email: "mymail2@foo.bar",
+        resource_type: "business_concept",
+        last_consumed_event: DateTime.utc_now(),
+        resource_id: @bc_4["id"]
+      }
+
+      [subscription_1, subscription_2]
+    end
   end
 
   defp create_users_in_cache do
@@ -152,6 +198,16 @@ defmodule TdAudit.NotificationDispatcherTest do
     |> Enum.map(&ConceptCache.put/1)
   end
 
+  defp rules_in_cache do
+    Enum.map(@rules, &RuleCache.put/1)
+  end
+
+  defp rule_results_in_cache do
+    resuls = Enum.map(@results, &RuleResultCache.put/1)
+    RuleResultCache.update_failed_ids(Enum.map(@results, & &1.id))
+    {:ok, resuls}
+  end
+
   defp atomize_keys(map) do
     for {key, val} <- map, into: %{}, do: {String.to_atom(key), val}
   end
@@ -161,23 +217,32 @@ defmodule TdAudit.NotificationDispatcherTest do
     |> Enum.map(&Audit.create_event(&1))
   end
 
-  defp subscriptions_fixture do
-    list_of_subscriptions()
+  defp subscriptions_fixture(event) do
+    event
+    |> list_of_subscriptions()
     |> Enum.map(&Subscriptions.create_subscription(&1))
     |> Enum.map(fn {:ok, value} -> value end)
   end
 
-  defp prepare_cache do
+  defp prepare_cache(:email_on_comment) do
     create_users_in_cache()
     create_bcs_in_cache()
   end
 
-  defp dispatch_notification_fixture do
-    {events_fixture(), subscriptions_fixture()}
+  defp prepare_cache(:failed_rule_results) do
+    create_users_in_cache()
+    create_bcs_in_cache()
+    rules_in_cache()
+    rule_results_in_cache()
   end
 
-  test "dispatch_notification/0" do
-    prepare_cache()
+  defp dispatch_notification_fixture do
+    {events_fixture(), subscriptions_fixture(:email_on_comment)}
+  end
+
+  test "dispatch_notification/1 on create comment" do
+    prepare_cache(:email_on_comment)
+    consumed_subscriptions = [1, 4]
     {_events_list, subs_list} = dispatch_notification_fixture()
     to_format_resource_id_1 = [nil: "mymail1@foo.bar", nil: "mymail2@foo.bar"]
     to_format_resource_id_2 = [nil: "mymail3@foo.bar"]
@@ -189,15 +254,36 @@ defmodule TdAudit.NotificationDispatcherTest do
 
     assert length(list_sent_notifications) == 2
 
-    Enum.any?(list_sent_notifications, fn el -> Map.fetch!(el, :to) == to_format_resource_id_1 end)
+    assert Enum.any?(list_sent_notifications, fn el ->
+             Map.fetch!(el, :to) == to_format_resource_id_1
+           end)
 
-    Enum.any?(list_sent_notifications, fn el -> Map.fetch!(el, :to) == to_format_resource_id_2 end)
+    assert Enum.any?(list_sent_notifications, fn el ->
+             Map.fetch!(el, :to) == to_format_resource_id_2
+           end)
 
     existing_subs = Subscriptions.list_subscriptions()
 
-    Enum.all?(existing_subs, fn sb ->
-      prior_sub = Enum.find(subs_list, &(sb.id == &1.id))
-      prior_sub.last_consumed_event < sb.last_consumed_event
-    end)
+    assert existing_subs
+           |> Enum.filter(&(Map.get(&1, :resource_id) in consumed_subscriptions))
+           |> Enum.all?(fn sb ->
+             prior_sub = Enum.find(subs_list, &(sb.id == &1.id))
+             prior_sub.last_consumed_event < sb.last_consumed_event
+           end)
+  end
+
+  test "dispatch_notification/1 on failed rule results" do
+    prepare_cache(:failed_rule_results)
+    subscriptions_fixture(:failed_rule_results)
+
+    expected_mails = [[nil: "mymail1@foo.bar"], [nil: "mymail2@foo.bar"]]
+
+    sent_notifications =
+      NotificationDispatcher.dispatch_notification(
+        {:dispatch_on_failed_results, "failed_rule_results"}
+      )
+
+    assert length(sent_notifications) == 2
+    assert Enum.all?(sent_notifications, &(Map.get(&1, :to) in expected_mails))
   end
 end